feat(mobile): preview the attached bill and smooth the picker-sheet drag
Tapping the receipt thumbnail on the capture screen now opens the bill
full-screen — pinch to zoom, drag to pan, double-tap to reset — the same
viewer the saved receipt uses. Extracted that viewer (ZoomableImage) into
a shared component so the two screens stay identical, and marked the
thumbnail with an expand glyph so its tappability reads.

Also refined the picker-sheet handle drag: it engages only after a clear
vertical move (a tap still closes), follows the finger down 1:1 while
rubber-banding an upward pull, carries a flick through to dismissal, and
springs back otherwise — so the currency and group sheets feel native to
pull down.
