+++ apps/mobile/src/app/receipt/[id].tsx
@@ -1,13 +1,11 @@
 import { useEffect, useState } from 'react';
 import Ionicons from '@expo/vector-icons/Ionicons';
-import { Image } from 'expo-image';
-import { Linking, useWindowDimensions, View } from 'react-native';
-import { Gesture, GestureDetector } from 'react-native-gesture-handler';
-import Animated, { useAnimatedStyle, useSharedValue, withTiming } from 'react-native-reanimated';
+import { Linking, View } from 'react-native';
 import { router, useLocalSearchParams } from 'expo-router';
 
 import { Button, EmptyState, IconButton, iconSize, Screen, useTheme } from '@waves/ui';
 
+import { ZoomableImage } from '@/components/ZoomableImage';
 import { fill, useStrings } from '@/i18n';
 import { providerFor } from '@/lib/cloud/providers';
 import { entryFor } from '@/lib/receiptIndex';
@@ -89,88 +87,5 @@ export default function ReceiptViewerScreen(): React.JSX.Element {
         )}
       </View>
     </Screen>
-  );
-}
-
-const AnimatedImage = Animated.createAnimatedComponent(Image);
-
-/**
- * The bill, pinch to zoom and drag to pan, double-tap to reset.
- *
- * Deliberately self-contained: the gestures live on shared values so the pan and
- * zoom stay on the UI thread, and the whole thing resets to fit on a double tap
- * so a person who zoomed in can always get back out. Clamped to a sane range so
- * the image can neither shrink to a dot nor fly off screen.
- */
-function ZoomableImage({ uri }: { uri: string }): React.JSX.Element {
-  const { width, height } = useWindowDimensions();
-
-  const scale = useSharedValue(1);
-  const savedScale = useSharedValue(1);
-  const translateX = useSharedValue(0);
-  const translateY = useSharedValue(0);
-  const savedX = useSharedValue(0);
-  const savedY = useSharedValue(0);
-
-  const pinch = Gesture.Pinch()
-    .onUpdate((event) => {
-      // Clamp between fit (1) and 5×, so it can neither vanish nor over-magnify.
-      const next = Math.min(Math.max(savedScale.get() * event.scale, 1), 5);
-      scale.set(next);
-    })
-    .onEnd(() => {
-      savedScale.set(scale.get());
-      if (scale.get() <= 1) {
-        translateX.set(withTiming(0));
-        translateY.set(withTiming(0));
-        savedX.set(0);
-        savedY.set(0);
-      }
-    });
-
-  const pan = Gesture.Pan()
-    .onUpdate((event) => {
-      // Panning only makes sense once zoomed in; at fit it stays put.
-      if (scale.get() <= 1) return;
-      translateX.set(savedX.get() + event.translationX);
-      translateY.set(savedY.get() + event.translationY);
-    })
-    .onEnd(() => {
-      savedX.set(translateX.get());
-      savedY.set(translateY.get());
-    });
-
-  const doubleTap = Gesture.Tap()
-    .numberOfTaps(2)
-    .onEnd(() => {
-      scale.set(withTiming(1));
-      savedScale.set(1);
-      translateX.set(withTiming(0));
-      translateY.set(withTiming(0));
-      savedX.set(0);
-      savedY.set(0);
-    });
-
-  const composed = Gesture.Simultaneous(pinch, pan, doubleTap);
-
-  const animatedStyle = useAnimatedStyle(() => ({
-    transform: [
-      { translateX: translateX.get() },
-      { translateY: translateY.get() },
-      { scale: scale.get() },
-    ],
-  }));
-
-  return (
-    <GestureDetector gesture={composed}>
-      <Animated.View style={{ flex: 1, justifyContent: 'center' }} collapsable={false}>
-        <AnimatedImage
-          source={{ uri }}
-          style={[{ width, height: height * 0.8 }, animatedStyle]}
-          contentFit="contain"
-          transition={150}
-        />
-      </Animated.View>
-    </GestureDetector>
   );
 }
